moved app routes enum elsewhere
